mintnft/requirements.txt
0x-contract-addresses==3.0.0
0x-contract-artifacts==3.0.0
0x-json-schemas==2.1.0
0x-order-utils==4.0.1
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
cytoolz==0.11.2
dacite==1.6.0
Deprecated==1.2.13
Django==4.0.4
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.3.0
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
Pillow==9.1.0
protobuf==3.20.0
pycodestyle==2.8.0
pycryptodome==3.14.1
pymerkle==2.0.2
pyrsistent==0.18.1
python-dotenv==0.20.0
requests==2.27.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.2
stringcase==1.2.0
thirdweb-contract-wrappers==2.0.4
thirdweb-eth-account==0.6.6
thirdweb-sdk==2.0.4
toml==0.10.2
toolz==0.11.2
tqdm==4.64.0
urllib3==1.26.9
varint==1.0.2
web3==5.27.0
websockets==9.1
wrapt==1.14.0
yarl==1.7.2
